Emitter improvements

+++ Emitters/ProjEmitter.cs
@@ -8,7 +8,7 @@ class ProjEmitter : ModHandlerEmitter<Projectile>
 {
 	public override Projectile Read(BinaryReader reader)
 	{
-		var proj =  Main.projectile[reader.ReadInt32()];
+		var proj =  Main.projectile[reader.ReadInt16()];
 		if(Main.netMode != NetmodeID.SinglePlayer &&  proj.owner == Main.myPlayer)
 		{
 			proj.netUpdate = true;
@@ -16,5 +16,5 @@ class ProjEmitter : ModHandlerEmitter<Projectile>
 		return proj;
 	}
 
-	public override void Write(BinaryWriter writer, Projectile ins) => writer.Write(ins.whoAmI);
+	public override void Write(BinaryWriter writer, Projectile ins) => writer.Write((short)ins.whoAmI);
 }
